Completed logic for all functions
Updated the divide() function to handle division by zero.

Made updates to displayTotal() to display an error if total is not a number.

Refactored the code for the 'equal' button action to handle when the
button was pressed when all numbers were not entered.

Modified the 'percent' button action to make correction calculations whether
start a new calculation or calculating from a previous calculation.